requirements.txt
Django>=4.2,<5.0
paho-mqtt>=1.6.1
pandas>=2.0.0
openpyxl>=3.0.0
django-crispy-forms>=1.14.0